Import `cn()` from `@/lib/utils`, Rename `defaultSize` to `defaultIconSize`
