+++ src/components/ui/svg/allesauto-logo.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@/utils';
+import { cn } from '@/lib/utils';
 
 export default function AllesautoLogo({
   className,
