+++ src/components/ui/svg/craft.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@/utils';
+import { cn } from '@/lib/utils';
 
 export default function Craft({ className }: { className?: string }) {
   return (
