+++ src/components/ui/svg/hdplus-logo.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@/utils';
+import { cn } from '@/lib/utils';
 
 export default function HdplusLogo({
   className,
