+++ src/components/ui/svg/miro.tsx
@@ -1,4 +1,4 @@
-import { defaultSize } from '@/components/ui/svg/config';
+import { defaultIconSize } from '@/components/ui/svg/config';
 
 export default function Miro({ className }: { className?: string }) {
   return (
@@ -6,7 +6,7 @@ export default function Miro({ className }: { className?: string }) {
       viewBox="0 0 128 128"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
-      className={className ? className : defaultSize}
+      className={className ? className : defaultIconSize}
     >
       <path
         d="M27.7333 0H100.267C115.467 0 128 12.5333 128 27.7333V100.267C128 115.467 115.467 128 100.267 128H27.7333C12.5333 128 0 115.467 0 100.267V27.7333C0 12.5333 12.5333 0 27.7333 0Z"
