--- src/components/ui/svg/tailwind-css.tsx
@@ -1,40 +1,0 @@
-export default function TailwindCSS({ className }: { className?: string }) {
-  return (
-    <svg
-      viewBox="0 0 128 79"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      className="h-20 w-auto"
-    >
-      <g>
-        <g id="Tailwind">
-          <g id="Clip path group">
-            <mask
-              id="mask0_748_1013"
-              maskUnits="userSpaceOnUse"
-              x="0"
-              y="0"
-              width="128"
-              height="77"
-            >
-              <g id="prefix__clip0">
-                <path id="Vector" d="M0 0H128V76.8H0V0Z" fill="white" />
-              </g>
-            </mask>
-            <g mask="url(#mask0_748_1013)">
-              <g id="Group">
-                <path
-                  id="Vector_2"
-                  fillRule="evenodd"
-                  clipRule="evenodd"
-                  d="M64 0C46.9333 0 36.2667 8.53333 32 25.6C38.4 17.0667 45.8667 13.8667 54.4 16C59.2687 17.216 62.7484 20.7502 66.6003 24.659C72.8747 31.0281 80.1375 38.4 96 38.4C113.067 38.4 123.733 29.8667 128 12.8C121.6 21.3333 114.133 24.5333 105.6 22.4C100.731 21.184 97.2516 17.6498 93.3997 13.741C87.1253 7.37185 79.8625 0 64 0ZM32 38.4C14.9333 38.4 4.26667 46.9333 0 64C6.4 55.4667 13.8667 52.2667 22.4 54.4C27.2687 55.6184 30.7484 59.1502 34.6003 63.059C40.8747 69.4282 48.1375 76.8 64 76.8C81.0667 76.8 91.7333 68.2667 96 51.2C89.6 59.7333 82.1333 62.9333 73.6 60.8C68.7313 59.584 65.2516 56.0498 61.3997 52.141C55.1253 45.7719 47.8625 38.4 32 38.4Z"
-                  fill="#38BDF8"
-                />
-              </g>
-            </g>
-          </g>
-        </g>
-      </g>
-    </svg>
-  );
-}
